refactor(*): addressed pr comments
