+++ services/workflows-service/src/business-report/dto/create-business-report.dto.ts
@@ -12,7 +12,7 @@ export class CreateBusinessReportDto {
   @IsOptional()
   @MinLength(1)
   @IsString()
-  businessId?: string;
+  businessCorrelationId?: string;
 
   @ApiProperty({
     required: true,
