+++ services/workflows-service/src/business-report/dtos/business-report-hook-search-query-params.dto.ts
@@ -1,0 +1,12 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsString, MinLength } from 'class-validator';
+
+export class BusinessReportHookSearchQueryParamsDto {
+  @ApiProperty({
+    required: true,
+    type: String,
+  })
+  @IsString()
+  @MinLength(1)
+  businessCorrelationId!: string;
+}
