+++ services/workflows-service/src/business/business.service.ts
@@ -46,8 +46,11 @@ export class BusinessService {
     return await this.repository.findById(id, args, projectIds);
   }
 
-  async getByIdUnscoped(id: string, args: Parameters<BusinessRepository['findByIdUnscoped']>[1]) {
-    return await this.repository.findByIdUnscoped(id, args);
+  async getByCorrelationIdUnscoped(
+    id: string,
+    args: Parameters<BusinessRepository['findByCorrelationIdUnscoped']>[1],
+  ) {
+    return await this.repository.findByCorrelationIdUnscoped(id, args);
   }
 
   async getByCorrelationId(correlationId: string, projectids: TProjectIds) {
